+++ jdaviz/configs/default/plugins/subset_tools/subset_tools.vue
@@ -45,6 +45,7 @@
           :show_if_single_entry="true"
           :mode.sync="subset_select_mode"
           :edit_value.sync="subset_edit_value"
+          :rename_error_message="rename_error_message"
           label="Subset"
           api_hint="plg.subset ="
           api_hint_rename="plg.rename_subset"