+++ jdaviz/core/loaders/importers/catalog/catalog.py
@@ -1,14 +1,15 @@
 from astropy.coordinates import SkyCoord
+from astropy.io.fits import BinTableHDU, HDUList, TableHDU
 from astropy.table import Table, QTable
 import astropy.units as u
 import numpy as np
-from traitlets import Bool, List, Unicode, observe
+from traitlets import Any, Bool, List, Unicode, observe
 
 from jdaviz.core.loaders.importers import BaseImporterToDataCollection
-from jdaviz.core.template_mixin import SelectPluginComponent
+from jdaviz.core.template_mixin import SelectFileExtensionComponent, SelectPluginComponent
 from jdaviz.core.registries import loader_importer_registry
 from jdaviz.core.user_api import ImporterUserApi
-from jdaviz.utils import RA_COMPS, DEC_COMPS
+from jdaviz.utils import RA_COMPS, DEC_COMPS, create_data_hash
 
 __all__ = ['CatalogImporter']
 
@@ -47,11 +48,46 @@ class CatalogImporter(BaseImporterToDataCollection):
     col_other_selected = List().tag(sync=True)
     col_other_multiselect = Bool(True).tag(sync=True)
 
+    # HDUList-specific options
+    input_has_extensions = Bool(False).tag(sync=True)
+    extension_items = List().tag(sync=True)
+    extension_selected = Any().tag(sync=True)
+    extension_multiselect = Bool(True).tag(sync=True)
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
         if not self.is_valid:
             return
+
+        if isinstance(self.input, HDUList):
+            self.input_has_extensions = True
+
+            ext_options = [{'label': f"{index}: [{hdu.name},{hdu.ver}]",
+                            'name': hdu.name,
+                            'ver': hdu.ver,
+                            'name_ver': f"{hdu.name},{hdu.ver}",
+                            'index': index,
+                            'data_hash': create_data_hash(hdu),
+                            'obj': hdu} for index, hdu in enumerate(self.input)]
+
+            self.extension = SelectFileExtensionComponent(self,
+                                                          items='extension_items',
+                                                          selected='extension_selected',
+                                                          multiselect='extension_multiselect',
+                                                          manual_options=ext_options,
+                                                          filters=[_validate_fits_tablehdu])
+
+            # the choices have already been filtered to only valid table HDUs, so
+            # choose the 0th to select the first valid table HDU by default
+            self.extension.selected = [self.extension.choices[0]]
+
+        else:
+            # otherwise, we got a Table object as input and can proceed
+            # to create the column selection components directly
+            self.input_has_extensions = False
+
+        input = self.input_as_table
 
         # dropdowns for catalogs with source positions in sky coordinates
         self.col_ra = SelectPluginComponent(self,
@@ -76,7 +112,7 @@ class CatalogImporter(BaseImporterToDataCollection):
         self.col_id = SelectPluginComponent(self,
                                             items='col_id_items',
                                             selected='col_id_selected',
-                                            manual_options=['Default (index)'] + self.input.colnames)  # noqa
+                                            manual_options=['Default (index)'] + input.colnames)  # noqa
 
         # dropdowns for tables with pixel source positions
         self.col_x = SelectPluginComponent(self,
@@ -92,8 +128,30 @@ class CatalogImporter(BaseImporterToDataCollection):
         self.col_other = SelectPluginComponent(self,
                                                items='col_other_items',
                                                selected='col_other_selected',
-                                               manual_options=self.input.colnames,
+                                               manual_options=input.colnames,
                                                multiselect='col_other_multiselect')
+
+    @property
+    def input_as_table(self):
+        if hasattr(self, 'extension'):
+            table_ext = self.extension.selected_obj[0]
+            if isinstance(table_ext, (TableHDU, BinTableHDU)):
+                return QTable(table_ext.data)
+        return self.input
+
+    @property
+    def is_valid(self):
+        if self.app.config not in ('deconfigged', 'imviz', 'mastviz'):
+            # NOTE: temporary during deconfig process
+            return False
+        if isinstance(self.input, (Table, QTable)) and len(self.input):
+            return True
+        elif isinstance(self.input, HDUList):
+            # check for the presence of at least one TableHDU/BinTableHDU extension
+            for i, hdu in enumerate(self.input):
+                if isinstance(hdu, (TableHDU, BinTableHDU)) and len(hdu.data) > 0:
+                    return True
+        return False
 
     def _guess_coord_cols(self, col):
         """
@@ -106,15 +164,16 @@ class CatalogImporter(BaseImporterToDataCollection):
         columns will be '---' (no selection)
         """
 
-        tab = self.input
-        colnames = self.input.colnames
+        input = self.input_as_table
+
+        colnames = input.colnames
 
         if colnames is None:
             return
 
         idx = None
         if col in ['ra', 'dec']:
-            col_is_sc = [isinstance(tab[colnames[i]], SkyCoord) for i in range(len(colnames))]
+            col_is_sc = [isinstance(input[colnames[i]], SkyCoord) for i in range(len(colnames))]
             if np.any(col_is_sc):
                 idx = np.where(col_is_sc)[0][0]
 
@@ -186,6 +245,8 @@ class CatalogImporter(BaseImporterToDataCollection):
 
         import_disabled = False
 
+        input = self.input_as_table
+
         if msg['name'] in ('col_ra_selected', 'col_dec_selected'):
 
             axis = ra if msg['name'] == 'col_ra_selected' else dec
@@ -202,13 +263,13 @@ class CatalogImporter(BaseImporterToDataCollection):
                 return
 
             has_units = False
-            if isinstance(self.input[axis], SkyCoord):
+            if isinstance(input[axis], SkyCoord):
                 has_units = True
-            elif hasattr(self.input[axis], 'unit'):
-                if self.input[axis].unit is not None:
+            elif hasattr(input[axis], 'unit'):
+                if input[axis].unit is not None:
                     has_units = True
                     # unit must be an angle unit
-                    if self.input[axis].unit.physical_type != 'angle':
+                    if input[axis].unit.physical_type != 'angle':
                         has_units = False
 
             # set the 'has units' traitlets for ra/dec, which determine if the unit
@@ -220,7 +281,7 @@ class CatalogImporter(BaseImporterToDataCollection):
 
             # disable import if the same ra and dec columns are selected
             # and they are NOT a SkyCoord column (which contains both RA and Dec),
-            if ra == dec and not isinstance(self.input[axis], SkyCoord):
+            if ra == dec and not isinstance(input[axis], SkyCoord):
                 import_disabled = True
             else:
                 import_disabled = False
@@ -248,19 +309,14 @@ class CatalogImporter(BaseImporterToDataCollection):
     @property
     def user_api(self):
         expose = ['col_ra', 'col_dec', 'col_x', 'col_y', 'col_id', 'col_other']
+        if self.input_has_extensions:
+            expose += ['extension']
         return ImporterUserApi(self, expose=expose)
 
     @property
-    def is_valid(self):
-        if self.app.config not in ('deconfigged', 'imviz', 'mastviz'):
-            # NOTE: temporary during deconfig process
-            return False
-        if isinstance(self.input, (Table, QTable)) and len(self.input):
-            return True
-        return False
-
-    @property
     def output_cols(self):
+
+        input = self.input_as_table
 
         coordinate_cols = []
         for col in [self.col_ra_selected, self.col_dec_selected]:
@@ -274,18 +330,23 @@ class CatalogImporter(BaseImporterToDataCollection):
 
         cols_all = coordinate_cols + self.col_other_selected
 
-        return [col for col in set(cols_all) if col in self.input.colnames]
+        return [col for col in set(cols_all) if col in input.colnames]
 
     @property
     def output(self):
 
-        table = self.input[self.output_cols]
+        input = self.input_as_table
+
+        if not isinstance(input, (Table, QTable)):
+            return
+
+        table = input[self.output_cols]
         output_table = QTable()
 
         # Handle RA / Dec columns, if selected
         if (self.col_ra_selected in table.colnames) and (self.col_dec_selected in table.colnames):  # noqa
-            ra = self.input[self.col_ra_selected]
-            dec = self.input[self.col_dec_selected]
+            ra = input[self.col_ra_selected]
+            dec = input[self.col_dec_selected]
 
             # The only modification made to the output table is the addition of individual
             # RA, Dec columns if the input columns are SkyCoord. This avoids unpacking
@@ -330,7 +391,7 @@ class CatalogImporter(BaseImporterToDataCollection):
         # handle output construction for X and Y coordinate columns, if selected
         if (self.col_x_selected in table.colnames) and (self.col_y_selected in table.colnames):  # noqa
             # if input is a string, try to convert to floats
-            if isinstance(self.input[self.col_x_selected][0], str):
+            if isinstance(input[self.col_x_selected][0], str):
                 try:
                     x_col = [float(x) for x in table[self.col_x_selected]]
                     output_table['X'] = x_col
@@ -339,7 +400,7 @@ class CatalogImporter(BaseImporterToDataCollection):
             else:
                 output_table['X'] = table[self.col_x_selected].astype(float)
 
-            if isinstance(self.input[self.col_y_selected][0], str):
+            if isinstance(input[self.col_y_selected][0], str):
                 try:
                     y_col = [float(y) for y in table[self.col_y_selected]]
                     output_table['Y'] = y_col
@@ -372,3 +433,8 @@ class CatalogImporter(BaseImporterToDataCollection):
                 output_table[col] = table[col]
 
         return output_table
+
+
+def _validate_fits_tablehdu(item):
+    hdu = item.get('obj')
+    return isinstance(hdu, (TableHDU, BinTableHDU))